+++ packages/adminui-vue/components/controls/UTable/formatValueMixin.js
@@ -1,10 +1,15 @@
+const lookups = require('@unitybase/adminui-vue/utils/lookups')
+
 module.exports = {
   methods: {
     formatValue ({ value, column, row }) {
       if (typeof column.format === 'function') {
         return column.format({ value, column, row })
       } else {
-        return value
+        const entity = column.attribute && column.attribute.associatedEntity
+        return entity
+          ? lookups.getValueById(entity, value)
+          : value
       }
     }
   }